Evaluation/requirements.txt
# Azure AI Evaluation requirements
azure-ai-evaluation>=1.0.0
azure-identity>=1.15.0
azure-ai-projects>=1.0.0

# Core dependencies
wikipedia>=1.4.0
python-dotenv>=1.0.0

# Optional: For enhanced functionality
openai>=1.0.0
requests>=2.28.0
